Movement fixes

+++ Assets/Code_MapOfOrders/Logic/MOMouseInputData.cs
@@ -5,7 +5,7 @@ namespace Code_MapOfOrders.Logic {
     [System.Serializable]
     public struct MOMouseInputData {
         public MouseAction mouseAction;
-        public float scrollValue;
+        public int scrollValue;
         public Vector2 pointerPosition;
     }
 }
